Create DeathsPlot.py
Creates a plot of Covid deaths for a single or for multiple counties/parishes or states. Can plot total deaths or as a function of population (for per capita comparisons). Program will create a local folder called "Data" where the .csv will download to. The Covid data is taken from https://github.com/datasets/covid-19 so thanks to the contributors for that.